Move c-authorship file component into its own file
